fefas: fix neighbor coarsening bug
The local logic to compute the size of a neighbor subdomain did not take
into account the dependence of the fine distribution on the coarse grid
layout.  Including that analytically would be complicated, so just send
messages with subdomain sizes (slightly more than necessary).  The new
test would create valgrind noise prior to the bug fix.