fefas: major refactor of coarse grids to balance each level
The previous method of coarsening selected processes for the coarse grid
based on process Z-order (Morton order), but left it up to the coarse
process set to decide how to partition the physical grid.  Once the
coarse set made a decision, the fine grids would be purely hierarchical.
This is an interesting optimization for vertical communication (which is
volumetric in size), but can create poor load balance.  The coarsest
grid is usually quite small and cannot be perfectly balanced on the
first refinement of a process set (up to 8 processes).  With the
strictly-hierarchical scheme, that load imbalance persists to all
children, and can be exacerbated by unfilled subdomains.

The new scheme balances each level independently, but process
representatives are chosen by Z-order so the messages will tend to stay
primarily local.  In this transformation, interpolation communicates
directly from global coarse space to local fine space.  In the previous
scheme, a horizontal neighbor communication was required to build a
coarse local space which was communicated to children.