fv: use "static inline" rather than pre-C99 GNU "inline"
gnu89-inline is a bad legacy feature.  C99 inline is providing an
optional definition to the compiler.  If the compiler chooses not to
inline (common with debugging options), then no definition is generated
and some other compilation unit is expected to generate the definition.
It appears that "static inline" semantics are desired in this case, so
use that.