FE: coarsen as needed to ensure coarsest grid has comm size=1
The decision of when to coarsen was previously defined in terms of
cmax, and tuned for process grids with prime factors less than 5.
This occasionally led to reaching an irreducible coarse problem size
while still on more than 1 process, which triggered an
error (otherwise poor performance could be expected).  Now we use cmax
as advisory to optionally coarsen earlier, but always coarsen the
process grid when necessary, even if it would lead to coarse or
intermediate levels that are larger than cmax.

Reported-by: Dmitry Duplyakin <dmitry.duplyakin@gmail.com>
Reported-by: Kit J. Menlove <Kit.Menlove@inl.gov>